requirements.txt
gradio==5.32.0
numpy>=2.3.0
pillow>=11.2.0
requests>=2.33.0
python-dotenv>=1.0.1
